# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "developer@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api-host.dev.app",
            "description": "測試區主機"
        },
        {
            "url": "https://api-host.production.app",
            "description": "正式區主機"
        },
        {
            "url": "http://localhost",
            "description": "Localhost"
        }
    ],
    "paths": {
        "/api/companies/types": {
            "get": {
                "tags": [
                    "company"
                ],
                "summary": "List all company types",
                "operationId": "",
                "parameters": [
                    {
                        "name": "petId",
                        "in": "path",
                        "description": "ID of pet to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "integer",
                                        "format": "int32"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "petstore_auth": [
                            "write:pets",
                            "read:pets"
                        ]
                    }
                ]
            }
        },
        "/api/misc/city_town": {
            "get": {
                "tags": [
                    "misc"
                ],
                "description": "取得縣市鄉鎮資料",
                "operationId": "cityTown",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "integer",
                                        "format": "int32"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "401": {
                        "description": "Not Authorized"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}